test for git repo and remove otherwise

+++ src/main/java/com/neuronrobotics/bowlerstudio/BowlerStudioMenuWorkspace.java
@@ -22,8 +22,9 @@ public class BowlerStudioMenuWorkspace {
 	private static Menu workspaceMenu;
 	private static final int maxMenueSize = 20;
 	private static boolean sorting = false;
-	private static HashMap<String,Integer> rank = new HashMap<String, Integer>();
+	private static HashMap<String, Integer> rank = new HashMap<String, Integer>();
 	private static boolean running = false;
+
 	public static void init(Menu workspacemenu) {
 		if (workspacemenu == null)
 			throw new RuntimeException();
@@ -31,39 +32,43 @@ public class BowlerStudioMenuWorkspace {
 	}
 
 	public static void loginEvent() {
-		if(running)
+		if (running)
 			return;
 		running = true;
 		rank.clear();
-		new Thread(()-> {
-			if(ScriptingEngine.hasNetwork())
-			for (int i=0;i<getWorkspaceData().keySet().size();i++) {
-				try {
-					String o = (String) getWorkspaceData().keySet().toArray()[i];
-					System.err.println("Pulling workspace "+o);
+		new Thread(() -> {
+			if (ScriptingEngine.hasNetwork())
+				for (int i = 0; i < getWorkspaceData().keySet().size(); i++) {
 					try {
-						if(!ScriptingEngine.isUrlAlreadyOpen(o))
-							ScriptingEngine.pull(o);
-					} catch (Throwable e) {
-						try {
-							ScriptingEngine.pull(o);
-						} catch (InvalidRemoteException ex) {
-							System.err.println("Deleting repo from workspace "+ex.getMessage());
+						String o = (String) getWorkspaceData().keySet().toArray()[i];
+						if (o.endsWith(".git")) {
+							System.err.println("Pulling workspace " + o);
+							try {
+								if (!ScriptingEngine.isUrlAlreadyOpen(o))
+									ScriptingEngine.pull(o);
+							} catch (Throwable e) {
+								try {
+									ScriptingEngine.pull(o);
+								} catch (InvalidRemoteException ex) {
+									System.err.println("Deleting repo from workspace " + ex.getMessage());
+									getWorkspaceData().remove(o);
+									ScriptingEngine.deleteRepo(o);
+								} catch (Throwable ex) {
+									e.printStackTrace();
+									getWorkspaceData().remove(o);
+									// ScriptingEngine.deleteRepo(o);
+									// i--;
+								}
+							}
+						}else {
 							getWorkspaceData().remove(o);
-							ScriptingEngine.deleteRepo(o);
-						}catch (Throwable ex) {
-							e.printStackTrace();
-							getWorkspaceData().remove(o);
-							//ScriptingEngine.deleteRepo(o);
-							//i--;
 						}
+					} catch (Exception e) {
+						e.printStackTrace();
 					}
-				} catch (Exception e) {
-					e.printStackTrace();
 				}
-			}
 			running = false;
-			
+
 		}).start();
 		sort();
 	}
@@ -74,24 +79,24 @@ public class BowlerStudioMenuWorkspace {
 
 	@SuppressWarnings("unchecked")
 	public static void add(String url, String menueMessage) {
-		if (menueMessage == null )
-			throw new RuntimeException("Menu Message can not be "+menueMessage);
-		if (menueMessage.length()<2) {
-			menueMessage= new Date().toString();
+		if (menueMessage == null)
+			throw new RuntimeException("Menu Message can not be " + menueMessage);
+		if (menueMessage.length() < 2) {
+			menueMessage = new Date().toString();
 		}
 		ArrayList<String> data;
-		if(getWorkspaceData()!=null)
-		synchronized (getWorkspaceData()) {
-			if (getWorkspaceData().get(url) == null) {
-				data = new ArrayList<String>();
-				data.add(menueMessage);
-				data.add(new Long(System.currentTimeMillis()).toString());
-				getWorkspaceData().put(url, data);
-				System.out.println("Workspace add: "+url);
+		if (getWorkspaceData() != null)
+			synchronized (getWorkspaceData()) {
+				if (getWorkspaceData().get(url) == null) {
+					data = new ArrayList<String>();
+					data.add(menueMessage);
+					data.add(new Long(System.currentTimeMillis()).toString());
+					getWorkspaceData().put(url, data);
+					System.out.println("Workspace add: " + url);
+				}
 			}
-		}
-		//data = (ArrayList<String>) workspaceData.get(url);
-		//data.set(1, new Long(System.currentTimeMillis()).toString());
+		// data = (ArrayList<String>) workspaceData.get(url);
+		// data.set(1, new Long(System.currentTimeMillis()).toString());
 		sort();
 		//
 
@@ -102,13 +107,13 @@ public class BowlerStudioMenuWorkspace {
 		if (sorting)
 			return;
 		sorting = true;
-		
-		boolean rankChanged=false;
+
+		boolean rankChanged = false;
 		try {
 			ArrayList<String> myOptions = new ArrayList<String>();
 			synchronized (getWorkspaceData()) {
 				for (String o : getWorkspaceData().keySet()) {
-					//System.out.println("Opt: "+o);
+					// System.out.println("Opt: "+o);
 					myOptions.add(o);
 				}
 			}
@@ -145,50 +150,46 @@ public class BowlerStudioMenuWorkspace {
 					remove(removedURL);
 				}
 			}
-			
-			for(int i=0;i<menu.size();i++) {
+
+			for (int i = 0; i < menu.size(); i++) {
 				String url = menu.get(i);
-				if(rank.get(url)==null) {
-					rankChanged=true;
-					rank.put(url,i);
-					//System.out.println("Rank firstNoted : "+url+" "+i);
+				if (rank.get(url) == null) {
+					rankChanged = true;
+					rank.put(url, i);
+					// System.out.println("Rank firstNoted : "+url+" "+i);
 				}
-				if(rank.get(url).intValue()!=i) {
-					rankChanged=true;
-					
+				if (rank.get(url).intValue() != i) {
+					rankChanged = true;
+
 				}
-				rank.put(url,i);
+				rank.put(url, i);
 			}
-			if(rankChanged) {
+			if (rankChanged) {
 				BowlerStudio.runLater(() -> {
 					if (workspaceMenu.getItems() != null)
 						workspaceMenu.getItems().clear();
-					
+
 					new Thread(() -> {
 						for (String url : menu) {
-							System.out.println("Workspace : "+url);
-								ArrayList<String> arrayList = (ArrayList<String>) getWorkspaceData().get(url);
-								if(arrayList!=null)
-									BowlerStudioMenu.setUpRepoMenue(workspaceMenu, 
-											url, 
-											false, 
-											false,
-											arrayList.get(0));
-							
+							System.out.println("Workspace : " + url);
+							ArrayList<String> arrayList = (ArrayList<String>) getWorkspaceData().get(url);
+							if (arrayList != null)
+								BowlerStudioMenu.setUpRepoMenue(workspaceMenu, url, false, false, arrayList.get(0));
+
 						}
 						sorting = false;
 					}).start();
 				});
-			}else {
+			} else {
 				sorting = false;
 			}
 
 		} catch (Exception ex) {
 			ex.printStackTrace();
 		}
-		if(rankChanged) {
+		if (rankChanged) {
 			System.out.println("Sorting workspace...");
-			new Thread(()->{
+			new Thread(() -> {
 				ConfigurationDatabase.save();
 			}).start();
 		}
@@ -203,6 +204,5 @@ public class BowlerStudioMenuWorkspace {
 			getWorkspaceData().remove(url);
 		}
 	}
-
 
 }
